+++ app/blueprints/main/controls.py
@@ -11,7 +11,6 @@ from flask import (
 
 from app.utils import *
 from app.models import GameServer
-from app.processes_global import *
 from app.forms.views import ValidateID, SendCommandForm, ServerControlForm, SelectCfgForm
 from app.services.controls_service import ControlService
 from app import cache