+++ app/blueprints/main/edit.py
@@ -10,7 +10,6 @@ from flask import (
 
 from app.utils import *
 from app.models import GameServer
-from app.processes_global import *
 from app.forms.views import UploadTextForm, DownloadCfgForm, SelectCfgForm
 
 from app.config.config_manager import ConfigManager